requirements.txt
crewai==0.1.20
google-generativeai==0.3.1
streamlit==1.31.0
pandas==2.1.4
scikit-learn==1.3.2
plotly==5.18.0
numpy==1.24.3
python-dotenv==1.0.0
